added a no-op ansible play call to Vagrantfile
  * When you use the ansible provisioner for Vagrant it will auto
create an inventory file for you including the appropriate mgmt IP
of the vm.  It makes things easier to rely on this file.  We run a
no-op playbook because there is already a shell based provisoner
script that runs that reboots the vms and renames the network
interfaces in a way that our ansible playbooks rely upon.  also the
reboot takes forever on cumulus vms so trying to auto run an ansible
playbook that actually changes configuration after the shell
provisioner runs just fails.