pinned cumulus-vx to 5.6.0 due to ssh issue to 5.7.0
  * cumulus-vx v5.7.0 was released recently, but when I try to
vagrant up a box with that version Vagrant fails to ssh to it early
on in the build process.  Last time this meant that something was
not matching with either the vagrant version and/or ssh ciphers
allowed by the vm.  I don't have time to investigate right now,
pinning to 5.6.0 fixes the issue