Turned off tls-proxy requirement for ovn-bgp-agent
  * This requirement makes multi-node devstack more challenging to
deploy.  I confirmed with agent auther Luis Tomas that there is not
really a requirement to have TLS enabled.  For now we have to clone
the repo and then make a changed to the lines that check for tls-
proxy.
  The force flag added to the git clone task is there because
otherwise you get an error when there are local changes to files
in the local staging area.  we don't care if these get discarded
because the lineinfile task will re-implement the changes.